Finished the operations + docs on the Customer part

+++ OOP_Project/OOP_Project/Server.cpp
@@ -19,6 +19,10 @@ Server::Server()
 	}
 }
 
+/// <summary>
+/// Functie pentru adaugarea unui vehicul de tip 'Masinia' in baza de date
+/// </summary>
+/// <param name="masina"></param>
 void Server::AddVehicle(Masina masina)
 {
 	try
@@ -45,6 +49,11 @@ void Server::AddVehicle(Masina masina)
 	}
 }
 
+
+/// <summary>
+/// Functie pentru adaugarea unui vehicul de tip 'Camion' in baza de date
+/// </summary>
+/// <param name="camion"></param>
 void Server::AddVehicle(Camion camion)
 {
 	con->setSchema("vehicles");
@@ -63,6 +72,11 @@ void Server::AddVehicle(Camion camion)
 	
 }
 
+
+/// <summary>
+/// Functie pentru adaugarea unui vehicul de tip 'Motocicleta' in baza de date
+/// </summary>
+/// <param name="motocicleta"></param>
 void Server::AddVehicle(Motocicleta motocicleta)
 {
 	con->setSchema("vehicles");
@@ -81,6 +95,13 @@ void Server::AddVehicle(Motocicleta motocicleta)
 	
 }
 
+
+/// <summary>
+/// Functie care cauta un id in tabela 'Customers' in functie de 'username' si 'parola'
+/// </summary>
+/// <param name="username"></param>
+/// <param name="password"></param>
+/// <returns></returns>
 int Server::GetCustomerId(string username, string password)
 {
 	try
@@ -102,6 +123,13 @@ int Server::GetCustomerId(string username, string password)
 	return -1;
 }
 
+
+/// <summary>
+/// Functie care cauta un id in tabela 'Admins' in functie de 'username' si 'parola'
+/// </summary>
+/// <param name="username"></param>
+/// <param name="password"></param>
+/// <returns></returns>
 int Server::GetAdminId(string username, string password)
 {
 	try
@@ -124,7 +152,10 @@ int Server::GetAdminId(string username, string password)
 }
 
 
-
+/// <summary>
+/// Functie care cauta si afiseaza pe ecran toate vehiculele detinute de 'client'
+/// </summary>
+/// <param name="client"></param>
 void Server::ShowOwnedVehicles(Client client)
 {
 	int id = GetCustomerId(client.GetUsername(), client.GetParola());
@@ -228,6 +259,11 @@ void Server::ShowOwnedVehicles(Client client)
 	}
 }
 
+
+/// <summary>
+/// Functie care cauta si afiseaza pe ecran toate vehiculele de tip 'Masina', detinute de 'client'
+/// </summary>
+/// <param name="client"></param>
 void Server::ShowVehicleTypeCar(Client client)
 {
 	int id = GetCustomerId(client.GetUsername(), client.GetParola());
@@ -266,6 +302,11 @@ void Server::ShowVehicleTypeCar(Client client)
 	}
 }
 
+
+/// <summary>
+/// Functie care cauta si afiseaza pe ecran toate vehiculele de tip 'Camion', detinute de 'client'
+/// </summary>
+/// <param name="client"></param>
 void Server::ShowVehicleTypeTruck(Client client)
 {
 	int id = GetCustomerId(client.GetUsername(), client.GetParola());
@@ -303,6 +344,11 @@ void Server::ShowVehicleTypeTruck(Client client)
 	}
 }
 
+
+/// <summary>
+/// Functie care cauta si afiseaza pe ecran toate vehiculele de tip 'Motocicleta', detinute de 'client'
+/// </summary>
+/// <param name="client"></param>
 void Server::ShowVehicleTypeMotorcycle(Client client)
 {
 	int id = GetCustomerId(client.GetUsername(), client.GetParola());
@@ -339,12 +385,19 @@ void Server::ShowVehicleTypeMotorcycle(Client client)
 	}
 }
 
-void Server::ShowAllVehicles()
+
+/// <summary>
+/// Functie care cauta si afiseaza pe ecran toate vehiculele care NU sunt detinute de 'client'
+/// </summary>
+/// <param name="client"></param>
+void Server::ShowAllVehicles(Client client)
 {
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
 	try
 	{
 		con->setSchema("vehicles");
-		pstmt = con->prepareStatement("SELECT * FROM vehicles.car");
+		pstmt = con->prepareStatement("SELECT * FROM vehicles.car WHERE OwnerId != ?");
+		pstmt->setInt(1,id);
 		auto result = pstmt->executeQuery();
 		auto data = result->next();
 		if (data) {
@@ -377,7 +430,8 @@ void Server::ShowAllVehicles()
 	try
 	{
 		con->setSchema("vehicles");
-		pstmt = con->prepareStatement("SELECT * FROM vehicles.motorcycle");
+		pstmt = con->prepareStatement("SELECT * FROM vehicles.motorcycle WHERE OwnerId != ?");
+		pstmt->setInt(1, id);
 		auto result = pstmt->executeQuery();
 		auto data = result->next();
 		if (data) {
@@ -410,7 +464,8 @@ void Server::ShowAllVehicles()
 	try
 	{
 		con->setSchema("vehicles");
-		pstmt = con->prepareStatement("SELECT * FROM vehicles.truck");
+		pstmt = con->prepareStatement("SELECT * FROM vehicles.truck WHERE OwnerId != ?");
+		pstmt->setInt(1, id);
 		auto result = pstmt->executeQuery();
 		auto data = result->next();
 		if (data) {
@@ -441,12 +496,144 @@ void Server::ShowAllVehicles()
 }
 
 
-void Server::UpdateCarById(int i, Masina masina)
+/// <summary>
+/// Functie care cauta si afiseaza pe ecran toate vehiculele de tip 'Masina', care NU sunt detinute de 'client'
+/// </summary>
+/// <param name="client"></param>
+void Server::ShowAllCars(Client client)
 {
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
 	try
 	{
 		con->setSchema("vehicles");
-		pstmt = con->prepareStatement("UPDATE vehicles.car SET Brand = ?, Model = ?, An = ?, KM = ?, Pret = ?, CapacitateCilindrica = ?, Combustibil = ?, Sasiu = ?, Transmisie = ?, NrUsi = ? WHERE VehicleID = ?");
+		pstmt = con->prepareStatement("SELECT * FROM vehicles.car WHERE OwnerId != ?");
+		pstmt->setInt(1, id);
+
+		auto result = pstmt->executeQuery();
+		auto data = result->next();
+		if (data) {
+			cout << "----------Masini----------\n";
+			while (data) {
+				int vehId = result->getInt("VehicleID");
+				string brand = result->getString("Brand");
+				string model = result->getString("Model");
+				int an = result->getInt("An");
+				int km = result->getInt("KM");
+				int pret = result->getInt("Pret");
+				int capCil = result->getInt("CapacitateCilindrica");
+				string combustibil = result->getString("Combustibil");
+				string sasiu = result->getString("Sasiu");
+				string transmisie = result->getString("Transmisie");
+				int usi = result->getInt("NrUsi");
+				cout << "\nMasina ID:" << vehId << "\n";
+				cout << Masina(Vehicul(brand, model, an, km, pret, capCil, combustibil, id), sasiu, transmisie, usi) << "\n\n";
+				data = result->next();
+			}
+			cout << "--------------------------\n";
+		}
+	}
+	catch (const std::exception& e)
+	{
+		cout << e.what() << "\n";
+	}
+}
+
+
+/// <summary>
+/// Functie care cauta si afiseaza pe ecran toate vehiculele de tip 'Camion', care NU sunt detinute de 'client'
+/// </summary>
+/// <param name="client"></param>
+void Server::ShowAllTrucks(Client client)
+{
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
+	try
+	{
+		con->setSchema("vehicles");
+		pstmt = con->prepareStatement("SELECT * FROM vehicles.truck WHERE OwnerId != ?");
+		pstmt->setInt(1, id);
+
+		auto result = pstmt->executeQuery();
+		auto data = result->next();
+		if (data) {
+			cout << "----------Camioane----------\n";
+			while (data) {
+				int vehId = result->getInt("VehicleID");
+				string brand = result->getString("Brand");
+				string model = result->getString("Model");
+				int an = result->getInt("An");
+				int km = result->getInt("KM");
+				int pret = result->getInt("Pret");
+				int capCil = result->getInt("CapacitateCilindrica");
+				string combustibil = result->getString("Combustibil");
+				int greutate = result->getInt("GreutateMaxima");
+				int lungime = result->getInt("LungimePlatforma");
+				cout << "\nCamion ID:" << vehId << "\n";
+				cout << Camion(Vehicul(brand, model, an, km, pret, capCil, combustibil, id), greutate, lungime) << "\n\n";
+				data = result->next();
+			}
+			cout << "----------------------------\n";
+		}
+	}
+	catch (const std::exception& e)
+	{
+		cout << e.what();
+	}
+}
+
+
+/// <summary>
+/// Functie care cauta si afiseaza pe ecran toate vehiculele de tip 'Motocicleta', care NU sunt detinute de 'client'
+/// </summary>
+/// <param name="client"></param>
+void Server::ShowAllMotorcycles(Client client)
+{
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
+	try
+	{
+		con->setSchema("vehicles");
+		pstmt = con->prepareStatement("SELECT * FROM vehicles.motorcycle WHERE OwnerId != ?");
+		pstmt->setInt(1, id);
+
+		auto result = pstmt->executeQuery();
+		auto data = result->next();
+		if (data) {
+			cout << "----------Motociclete----------\n";
+			while (data) {
+				string brand = result->getString("Brand");
+				string model = result->getString("Model");
+				int an = result->getInt("An");
+				int km = result->getInt("KM");
+				int pret = result->getInt("Pret");
+				int capCil = result->getInt("CapacitateCilindrica");
+				string combustibil = result->getString("Combustibil");
+				string motor = result->getString("TipMotor");
+				string corp = result->getString("Corp");
+				cout << Motocicleta(Vehicul(brand, model, an, km, pret, capCil, combustibil, id), motor, corp) << "\n";
+				data = result->next();
+			}
+			cout << "-------------------------------\n";
+		}
+
+	}
+	catch (const std::exception& e)
+	{
+		cout << e.what();
+	}
+}
+
+
+/// <summary>
+/// Functie care cauta si schimba un vehicul de tip 'Masina', dupa un id 'i'.
+/// Totodata, se verifica daca vehiculul este detinuta de catre 'client'
+/// </summary>
+/// <param name="client"></param>
+void Server::UpdateCarById(int i, Masina masina, Client client)
+{
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
+	try
+	{
+		con->setSchema("vehicles");
+		pstmt = con->prepareStatement("UPDATE vehicles.car SET Brand = ?, Model = ?, An = ?, KM = ?, Pret = ?, CapacitateCilindrica = ?, Combustibil = ?, Sasiu = ?, Transmisie = ?, NrUsi = ? WHERE VehicleID = ? AND OwnerId = ?");
 		pstmt->setString(1, masina.Getbrand());
 		pstmt->setString(2, masina.GetModel());
 		pstmt->setInt(3, masina.GetAn());
@@ -458,6 +645,7 @@ void Server::UpdateCarById(int i, Masina masina)
 		pstmt->setString(9, masina.GetTransmisie());
 		pstmt->setInt(10, masina.GetNrUsi());
 		pstmt->setInt(11, i);
+		pstmt->setInt(12, id);
 
 		auto result = pstmt->executeUpdate();
 	}
@@ -467,12 +655,19 @@ void Server::UpdateCarById(int i, Masina masina)
 	}
 }
 
-void Server::UpdateTruckById(int i, Camion camion)
+
+/// <summary>
+/// Functie care cauta si schimba un vehicul de tip 'Camion', dupa un id 'i'.
+/// Totodata, se verifica daca vehiculul este detinuta de catre 'client'
+/// </summary>
+/// <param name="client"></param>
+void Server::UpdateTruckById(int i, Camion camion, Client client)
 {
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
 	try
 	{
 		con->setSchema("vehicles");
-		pstmt = con->prepareStatement("UPDATE vehicles.truck SET Brand = ?, Model = ?, An = ?, KM = ?, Pret = ?, CapacitateCilindrica = ?, Combustibil = ?, GreutateMaxima = ?, LungimePlatforma = ? WHERE VehicleID = ?");
+		pstmt = con->prepareStatement("UPDATE vehicles.truck SET Brand = ?, Model = ?, An = ?, KM = ?, Pret = ?, CapacitateCilindrica = ?, Combustibil = ?, GreutateMaxima = ?, LungimePlatforma = ? WHERE VehicleID = ? AND OwnerId = ?");
 		pstmt->setString(1, camion.Getbrand());
 		pstmt->setString(2, camion.GetModel());
 		pstmt->setInt(3, camion.GetAn());
@@ -483,6 +678,7 @@ void Server::UpdateTruckById(int i, Camion camion)
 		pstmt->setInt(8, camion.GetGretutate());
 		pstmt->setInt(9, camion.GetLungime());
 		pstmt->setInt(10, i);
+		pstmt->setInt(11, id);
 
 		auto result = pstmt->executeUpdate();
 	}
@@ -492,12 +688,19 @@ void Server::UpdateTruckById(int i, Camion camion)
 	}
 }
 
-void Server::UpdateMotorcycleById(int i, Motocicleta motocicleta)
+
+/// <summary>
+/// Functie care cauta si schimba un vehicul de tip 'Motocicleta', dupa un id 'i'.
+/// Totodata, se verifica daca vehiculul este detinuta de catre 'client'
+/// </summary>
+/// <param name="client"></param>
+void Server::UpdateMotorcycleById(int i, Motocicleta motocicleta, Client client)
 {
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
 	try
 	{
 		con->setSchema("vehicles");
-		pstmt = con->prepareStatement("UPDATE vehicles.truck SET Brand = ?, Model = ?, An = ?, KM = ?, Pret = ?, CapacitateCilindrica = ?, Combustibil = ?, TipMotor = ?, Corp = ? WHERE VehicleID = ?");
+		pstmt = con->prepareStatement("UPDATE vehicles.truck SET Brand = ?, Model = ?, An = ?, KM = ?, Pret = ?, CapacitateCilindrica = ?, Combustibil = ?, TipMotor = ?, Corp = ? WHERE VehicleID = ? AND OwnerId = ?");
 		pstmt->setString(1, motocicleta.Getbrand());
 		pstmt->setString(2, motocicleta.GetModel());
 		pstmt->setInt(3, motocicleta.GetAn());
@@ -508,6 +711,7 @@ void Server::UpdateMotorcycleById(int i, Motocicleta motocicleta)
 		pstmt->setString(8, motocicleta.GetTipMotor());
 		pstmt->setString(9, motocicleta.GetCorp());
 		pstmt->setInt(10, i);
+		pstmt->setInt(11, id);
 
 		auto result = pstmt->executeUpdate();
 	}
@@ -518,14 +722,20 @@ void Server::UpdateMotorcycleById(int i, Motocicleta motocicleta)
 }
 
 
-
-void Server::DeleteCarById(int i)
+/// <summary>
+/// Functie care cauta si sterge un vehicul de tip 'Masina', dupa un id 'i'.
+/// Totodata, se verifica daca vehiculul este detinuta de catre 'client'
+/// </summary>
+/// <param name="client"></param>
+void Server::DeleteCarById(int i, Client client)
 {
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
 	try
 	{
 		con->setSchema("vehicles");
-		pstmt = con->prepareStatement("DELETE FROM vehicles.car WHERE VehicleId = ?");
+		pstmt = con->prepareStatement("DELETE FROM vehicles.car WHERE VehicleId = ? AND OwnerId = ? ");
 		pstmt->setInt(1, i);
+		pstmt->setInt(2, id);
 		pstmt->execute();
 	}
 	catch (const std::exception& e)
@@ -534,13 +744,21 @@ void Server::DeleteCarById(int i)
 	}
 }
 
-void Server::DeleteTruckById(int i)
+
+/// <summary>
+/// Functie care cauta si sterge un vehicul de tip 'Camion', dupa un id 'i'.
+/// Totodata, se verifica daca vehiculul este detinuta de catre 'client'
+/// </summary>
+/// <param name="client"></param>
+void Server::DeleteTruckById(int i, Client client)
 {
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
 	try
 	{
 		con->setSchema("vehicles");
-		pstmt = con->prepareStatement("DELETE FROM vehicles.truck WHERE VehicleId = ?");
+		pstmt = con->prepareStatement("DELETE FROM vehicles.truck WHERE VehicleId = ? AND OwnerId = ?");
 		pstmt->setInt(1, i);
+		pstmt->setInt(2, id);
 		pstmt->execute();
 	}
 	catch (const std::exception& e)
@@ -549,13 +767,21 @@ void Server::DeleteTruckById(int i)
 	}
 }
 
-void Server::DeleteMotorcycleById(int i)
+
+/// <summary>
+/// Functie care cauta si sterge un vehicul de tip 'Motocicleta', dupa un id 'i'.
+/// Totodata, se verifica daca vehiculul este detinuta de catre 'client'
+/// </summary>
+/// <param name="client"></param>
+void Server::DeleteMotorcycleById(int i, Client client)
 {
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
 	try
 	{
 		con->setSchema("vehicles");
-		pstmt = con->prepareStatement("DELETE FROM vehicles.motorcycle WHERE VehicleId = ?");
+		pstmt = con->prepareStatement("DELETE FROM vehicles.motorcycle WHERE VehicleId = ? AND OwnerId = ?");
 		pstmt->setInt(1, i);
+		pstmt->setInt(2, id);
 		pstmt->execute();
 	}
 	catch (const std::exception& e)
@@ -564,3 +790,162 @@ void Server::DeleteMotorcycleById(int i)
 	}
 }
 
+
+/// <summary>
+/// Functie care cauta si cumpara un vehicul de tip 'Masina', dupa un id 'i'.
+/// Totodata, se verifica daca vehiculul NU este detinuta de catre 'client'.
+/// Functia returneaza un buget modificat al obiectului 'client', daca vehiculul poate fi achizitionat. In caz contrat va returna bugetul actual fara modificari.
+/// </summary>
+/// <param name="client"></param>
+int Server::PurchaseCarById(int i, Client client)
+{
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
+
+	try
+	{
+		con->setSchema("vehicles");
+		pstmt = con->prepareStatement("SELECT * FROM vehicles.car WHERE VehicleID = ? AND OwnerId != ?");
+		pstmt->setInt(1, i);
+		pstmt->setInt(2, id);
+
+		auto result = pstmt->executeQuery();
+		auto data = result->next();
+		if (data) {
+			int vehPrice = result->getInt("Pret");
+
+			if (vehPrice > client.GetBuget()) {
+				cout << "Fonduri insuficiente\n";
+			}
+			else {
+				pstmt = con->prepareStatement("UPDATE vehicles.car SET OwnerId = ? WHERE VehicleID = ?");
+				pstmt->setInt(1, id);
+				pstmt->setInt(2, i);
+				pstmt->execute();
+
+				con->setSchema("users");
+				int updatedBuget = client.GetBuget() - vehPrice;
+
+				pstmt = con->prepareStatement("UPDATE users.customers SET Buget = ? WHERE ID = ?");
+				pstmt->setInt(1, updatedBuget);
+				pstmt->setInt(2, id);
+				pstmt->execute();
+				cout << "Masina cumparata\n";
+				return updatedBuget;
+			}
+		}
+		else {
+			cout << "Entry error\n";
+		}
+		return client.GetBuget();
+	}
+	catch (const std::exception& e)
+	{
+		cout << e.what() << "\n";
+	}
+}
+
+
+/// <summary>
+/// Functie care cauta si cumpara un vehicul de tip 'Camion', dupa un id 'i'.
+/// Totodata, se verifica daca vehiculul NU este detinuta de catre 'client'.
+/// Functia returneaza un buget modificat al obiectului 'client', daca vehiculul poate fi achizitionat. In caz contrat va returna bugetul actual fara modificari.
+/// </summary>
+/// <param name="client"></param>
+int Server::PurchaseTruckById(int i, Client client)
+{
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
+
+	try
+	{
+		con->setSchema("vehicles");
+		pstmt = con->prepareStatement("SELECT * FROM vehicles.truck WHERE VehicleID = ? AND OwnerId != ?");
+		pstmt->setInt(1, i);
+		pstmt->setInt(2, id);
+
+		auto result = pstmt->executeQuery();
+		auto data = result->next();
+		if (data) {
+			int vehPrice = result->getInt("Pret");
+
+			if (vehPrice > client.GetBuget()) {
+				cout << "Fonduri insuficiente";
+			}
+			else {
+				pstmt = con->prepareStatement("UPDATE vehicles.truck SET OwnerId = ? WHERE VehicleID = ?");
+				pstmt->setInt(1, id);
+				pstmt->setInt(2, i);
+				pstmt->execute();
+
+				con->setSchema("users");
+				int updatedBuget = client.GetBuget() - vehPrice;
+				pstmt = con->prepareStatement("UPDATE users.customers SET Buget = ? WHERE ID = ?");
+				pstmt->setInt(1, updatedBuget);
+				pstmt->setInt(2, id);
+				pstmt->execute();
+				cout << "Camion cumparat";
+				return updatedBuget;
+			}
+		}
+		else {
+			cout << "Entry error\n";
+		}
+		return client.GetBuget();
+	}
+	catch (const std::exception& e)
+	{
+		cout << e.what() << "\n";
+	}
+}
+
+
+/// <summary>
+/// Functie care cauta si cumpara un vehicul de tip 'Motocicleta', dupa un id 'i'.
+/// Totodata, se verifica daca vehiculul NU este detinuta de catre 'client'.
+/// Functia returneaza un buget modificat al obiectului 'client', daca vehiculul poate fi achizitionat. In caz contrat va returna bugetul actual fara modificari.
+/// </summary>
+/// <param name="client"></param>
+int Server::PurchaseMotorcycleById(int i, Client client)
+{
+	int id = GetCustomerId(client.GetUsername(), client.GetParola());
+
+	try
+	{
+		con->setSchema("vehicles");
+		pstmt = con->prepareStatement("SELECT * FROM vehicles.motorcycle WHERE VehicleID = ? AND OwnerId != ?");
+		pstmt->setInt(1, i);
+		pstmt->setInt(2, id);
+
+		auto result = pstmt->executeQuery();
+		auto data = result->next();
+		if (data) {
+			int vehPrice = result->getInt("Pret");
+
+			if (vehPrice > client.GetBuget()) {
+				cout << "Fonduri insuficiente";
+			}
+			else {
+				pstmt = con->prepareStatement("UPDATE vehicles.motorcycle SET OwnerId = ? WHERE VehicleID = ?");
+				pstmt->setInt(1, id);
+				pstmt->setInt(2, i);
+				pstmt->execute();
+
+				con->setSchema("users");
+				int updatedBuget = client.GetBuget() - vehPrice;
+				pstmt = con->prepareStatement("UPDATE users.customers SET Buget = ? WHERE ID = ?");
+				pstmt->setInt(1, updatedBuget);
+				pstmt->setInt(2, id);
+				pstmt->execute();
+				cout << "Motocicleta cumparata";
+				return updatedBuget;
+			}
+		}
+		else {
+			cout << "Entry error\n";
+		}
+		return client.GetBuget();
+	}
+	catch (const std::exception& e)
+	{
+		cout << e.what() << "\n";
+	}
+}
